#58 add file for GATTAltitude
